Rename some old 'java' directories to 'kotlin'
This is actually required for Spotless to target the Kotlin sources.
