+++ samples/emoji-search/android-views/src/main/kotlin/com/example/redwood/emojisearch/android/views/ViewImage.kt
